requirements.txt
streamlit
langchain
langchain-openai
chromadb
sentence-transformers
psycopg2-binary
pandas
numpy
python-dotenv
pypdf
langchain-text-splitters
langchain-community
pandasai
langchain_ollama
sqlalchemy 